register method added

+++ courseManagementSystem/src/main/java/project/courseManagementSystem/business/concretes/InstructorManager.java
@@ -6,7 +6,12 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import project.courseManagementSystem.business.abstracts.InstructorService;
+import project.courseManagementSystem.business.abstracts.UserService;
+import project.courseManagementSystem.business.validationRules.InstructorValidatorService;
+import project.courseManagementSystem.core.email.EmailCheckService;
 import project.courseManagementSystem.core.utilities.results.DataResult;
+import project.courseManagementSystem.core.utilities.results.ErrorDataResult;
+import project.courseManagementSystem.core.utilities.results.ErrorResult;
 import project.courseManagementSystem.core.utilities.results.Result;
 import project.courseManagementSystem.core.utilities.results.SuccessDataResult;
 import project.courseManagementSystem.core.utilities.results.SuccessResult;
@@ -17,13 +22,39 @@ import project.courseManagementSystem.entities.concretes.Instructor;
 public class InstructorManager implements InstructorService{
 
 	private InstructorDao instructorDao;
+	private InstructorValidatorService instructorValidatorService;
+	private EmailCheckService emailCheckService;
+	private UserService userService;
+	
 	
 	@Autowired
-	public InstructorManager(InstructorDao instructorDao) {
+	public InstructorManager(InstructorDao instructorDao, InstructorValidatorService instructorValidatorService,
+			EmailCheckService emailCheckService, UserService userService) {
 		super();
 		this.instructorDao = instructorDao;
+		this.instructorValidatorService = instructorValidatorService;
+		this.emailCheckService = emailCheckService;
+		this.userService = userService;
 	}
 
+	@Override
+	public Result register(Instructor instructor) {
+		if(!instructorValidatorService.checkIfInstructorInfoIsFull(instructor)) {
+			return new ErrorResult("enter all your information completely");
+		}
+		
+		if(!emailCheckService.emailCheck(instructor.getEmail())) {
+			return new ErrorResult("invalid email");
+		}
+		
+		if(userService.existsByEmail(instructor.getEmail())) {
+			return new ErrorResult("instructor is already exist");
+		}
+			
+		instructorDao.save(instructor);
+		return new SuccessResult("Successfully registered");
+	}
+	
 	@Override
 	public Result add(Instructor entity) {
 		instructorDao.save(entity);
@@ -32,29 +63,37 @@ public class InstructorManager implements InstructorService{
 
 	@Override
 	public Result delete(int id) {
-		instructorDao.deleteById(id);
-		return new SuccessResult("Instructor deleted!");
+		
+		if(userService.existById(id)) {
+			instructorDao.deleteById(id);
+			return new SuccessResult("Instructor deleted!");
+		}
+		
+		return new ErrorResult("instructor is not exist");	
 	}
 
 	@Override
 	public Result update(Instructor entity) {
-		// refactor this:
 		Instructor updatedUser = getById(entity.getId()).getData();
+		
 		updatedUser.setFirstName(entity.getFirstName());
 		updatedUser.setLastName(entity.getLastName());
 		updatedUser.setEmail(entity.getEmail());
 		updatedUser.setDepartmentName(entity.getDepartmentName());
 		updatedUser.setPhoneNumber(entity.getPhoneNumber());
 		updatedUser.setNationalityId(entity.getNationalityId());
-		//instructorDao.delete(entity);
-		// instructorDao.save(updatedUser);
+		
+		instructorDao.save(updatedUser);
 		return new SuccessResult("Instructor updated!");
 	}
 
 	@Override
 	public DataResult<Instructor> getById(int id) {
-		;
-		return new SuccessDataResult<Instructor>(instructorDao.getById(id), "Instructor viewed!");
+		Instructor instructor = instructorDao.findById(id);
+		if(instructor == null) {
+			return new ErrorDataResult<Instructor>(null, "Instructor is not exist");
+		}
+		return new SuccessDataResult<Instructor>(instructorDao.findById(id), "Instructor viewed!");
 	}
 
 	@Override
